api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
lxml>=6.0.2
anthropic>=0.42.0
pydantic>=2.10.0
python-dotenv>=1.0.0
playwright>=1.56.0
Pillow>=12.1.0
validators>=0.22.0
